+++ src/components/Event/EventModal/EventDetail/EventDetail.tsx
@@ -21,6 +21,7 @@ import { useProfileSettingsContext } from "../../../../context/ProfileSetting"
 import { SessionEventAttributes } from "../../../../entities/Event/types"
 import { profileSiteUrl } from "../../../../entities/Event/utils"
 import { canEditAnyEvent } from "../../../../entities/ProfileSettings/utils"
+import { useDecentralandFoundationAddresses } from "../../../../hooks/useDecentralandFoundationAddresses"
 import extraIcon from "../../../../images/info.svg"
 import friendsIcon from "../../../../images/secondary-friends.svg"
 import infoIcon from "../../../../images/secondary-info.svg"
@@ -36,7 +37,6 @@ import JumpInButton from "../../../Button/JumpInPosition"
 import MenuIcon, { MenuIconItem } from "../../../MenuIcon/MenuIcon"
 import EventSection from "../../EventSection"
 import EventDateDetail from "./EventDateDetail"
-import { useDecentralandFoundationAddresses } from "./utils"
 
 import "./EventDetail.css"
